+++ extensions/1bookmark/src/components/BookmarkItem.tsx
@@ -3,13 +3,16 @@ import { BookmarkItemActionPanel } from "./BookmarkItemActionPanel";
 import { RouterOutputs } from "../utils/trpc.util";
 import { getFavicon } from "@raycast/utils";
 import { useMemo } from "react";
+import { RankingDatas } from "../types";
 
 export const BookmarkItem = (props: {
   bookmark: RouterOutputs["bookmark"]["listAll"][number];
   me?: RouterOutputs["user"]["me"];
   refetch: () => void;
+  rankingDatas: RankingDatas;
+  setRankingDatas: (rankingDatas: RankingDatas | ((prev: RankingDatas) => RankingDatas)) => void;
 }) => {
-  const { bookmark, me, refetch } = props;
+  const { bookmark, me, refetch, setRankingDatas, rankingDatas } = props;
   const { name, url, spaceId, tags } = bookmark;
   const space = me?.associatedSpaces.find((s) => s.id === spaceId);
 
@@ -39,7 +42,14 @@ export const BookmarkItem = (props: {
         { icon: space?.image || (space?.type === "PERSONAL" ? Icon.Person : Icon.TwoPeople) },
       ]}
       actions={
-        <BookmarkItemActionPanel bookmark={bookmark} toggleBookmarkDetail={() => {}} me={me} refetch={refetch} />
+        <BookmarkItemActionPanel
+          bookmark={bookmark}
+          toggleBookmarkDetail={() => {}}
+          me={me}
+          refetch={refetch}
+          rankingDatas={rankingDatas}
+          setRankingDatas={setRankingDatas}
+        />
       }
     />
   );